Commented out new asset move stuff
