+++ engine/include/assets/AssetUtilities.h
