+++ engine/source/engine/Context.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "assets/Assets.h"
 #include "ecs/Registry.h"
 #include "math/Random.h"
 #include "time/Time.h"
@@ -11,5 +12,6 @@ namespace nc
         Registry registry;
         time::Time time;
         nc::Random random;
+        // nc::Assets assets;
     };
 }
